Reset cloud layers after minimisation
The minimisation routine LowWaterCloud.minimize_cbh calls
LowWaterCloud.init_cloud_layers, so in every round of the iterations the
cloud layers get reset.  We don't know if the final call to the
minimisation routine corresponds to the function minimum; for basin
hopping it might, for brute force it likely won't.  This was triggering
issue #29.  This commit re-initialises the layers with the final
minimisation value.

Fixes #29.