Set up travis and stickler for fogpy
Set up travis and stickler for fogpy.  The fogpy testing output is
very verbose and very slow.  When using test runners that do not
capture the output, the build failed due to the logfile exceeding
4 MB. When using a test runner that does capture the output,
the build fails due to not producing any output for 10 minutes.
See https://stackoverflow.com/q/56668791/974555

Therefore it is currently using a travis_wait 45 for the call to pytest
as a test runner.  I'm not entirely sure

Note that tests on travis currently fail even though tests locally
succeed.  For some failures I understand why, for others I'm not sure,
but see PR #42 for one attempt to resolve this.